+++ apps/AlbionMC_API/src/repository/data_source_repository.py
@@ -1,5 +1,6 @@
 from typing import List
 
+from fastapi import HTTPException
 from injector import inject
 
 from src.core.entities.data_source import DataSource
@@ -15,6 +16,12 @@ class DataSourceRepository:
         self.conn = ctx.conn
 
     def new(self, record: DataSource, user_name: str = "repository") -> None:
+        ds = self.get_by_name(record_name=record.name)
+        if(ds):
+            raise HTTPException(detail={
+                'id'      : ds.id,
+                'message' : 'Data Source with equal name exists...',
+            },status_code=400)
         with self.conn.cursor() as cur:
             query = "INSERT INTO data_sources (data_source_name, trust_level, updated_by, created_by) VALUES (%s, %s, %s, %s)"
             cur.execute(
@@ -33,6 +40,24 @@ class DataSourceRepository:
         with self.conn.cursor() as cur:
             query = "SELECT id, data_source_name, trust_level, updated_at, updated_by, created_at, created_by FROM data_sources WHERE id = %s"
             cur.execute(query, (record_id,))
+            row = cur.fetchone()
+            if row:
+                return DataSource(**{
+                    "id": row[0],
+                    "name": row[1],
+                    "trust_level": row[2],
+                    "updated_at": row[3],
+                    "updated_by": row[4],
+                    "created_at": row[5],
+                    "created_by": row[6]
+                })
+            else:
+                return None
+
+    def get_by_name(self, record_name: str) -> DataSource | None:
+        with self.conn.cursor() as cur:
+            query = "SELECT id, data_source_name, trust_level, updated_at, updated_by, created_at, created_by FROM data_sources WHERE data_source_name = %s"
+            cur.execute(query, (record_name,))
             row = cur.fetchone()
             if row:
                 return DataSource(**{
